+++ lib/volt/config/server.ex
@@ -4,6 +4,7 @@ defmodule Volt.Config.Server do
   defstruct prefix: Volt.Paths.prefix(),
             watch_dirs: [],
             reload_dirs: [],
+            watch_ignored: [],
             hmr_timeout: 60_000,
             watch: true
 end